+++ src/games/alpha-adjust/src/interfaces/crystal.interface.ts
@@ -1,0 +1,8 @@
+export interface ICrystalConstructor {
+  scene: Phaser.Scene;
+  x: number;
+  y: number;
+  texture: string;
+  alpha?: number;
+  frame?: string | number;
+}