+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Refactor `point-in-polygon`
 - Refactor `procedural-generation`
 - Refactor `raycasting`
+- Refactor `alpha-adjust`
 
 ---